move asyncio support into aio sub-module
